rename transformer_layer as transformer_decoder_layer
